+++ tests/classification/test_average_precision.py
@@ -40,9 +40,9 @@ def _sk_average_precision_score(y_true, probas_pred, num_classes=1, average=None
         y_true_temp[y_true == i] = 1
         res.append(sk_average_precision_score(y_true_temp, probas_pred[:, i]))
 
-    if average == 'macro':
+    if average == "macro":
         return np.array(res).mean()
-    elif average == 'weighted':
+    elif average == "weighted":
         weights = np.bincount(y_true) if y_true.max() > 1 else y_true.sum(axis=0)
         weights = weights / sum(weights)
         return (np.array(res) * weights).sum()
@@ -85,13 +85,13 @@ def _sk_avg_prec_multidim_multiclass_prob(preds, target, num_classes=1, average=
         (_input_multilabel.preds, _input_multilabel.target, _sk_avg_prec_multilabel_prob, NUM_CLASSES),
     ],
 )
-@pytest.mark.parametrize("average", ['micro', 'macro', 'weighted', None])
+@pytest.mark.parametrize("average", ["micro", "macro", "weighted", None])
 class TestAveragePrecision(MetricTester):
     @pytest.mark.parametrize("ddp", [True, False])
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
     def test_average_precision(self, preds, target, sk_metric, num_classes, average, ddp, dist_sync_on_step):
-        if target.max() > 1 and average == 'micro':
-            pytest.skip('average=micro and multiclass input cannot be used together')
+        if target.max() > 1 and average == "micro":
+            pytest.skip("average=micro and multiclass input cannot be used together")
 
         self.run_class_metric_test(
             ddp=ddp,
@@ -100,19 +100,19 @@ class TestAveragePrecision(MetricTester):
             metric_class=AveragePrecision,
             sk_metric=partial(sk_metric, num_classes=num_classes, average=average),
             dist_sync_on_step=dist_sync_on_step,
-            metric_args={"num_classes": num_classes, 'average': average},
+            metric_args={"num_classes": num_classes, "average": average},
         )
 
     def test_average_precision_functional(self, preds, target, sk_metric, num_classes, average):
-        if target.max() > 1 and average == 'micro':
-            pytest.skip('average=micro and multiclass input cannot be used together')
+        if target.max() > 1 and average == "micro":
+            pytest.skip("average=micro and multiclass input cannot be used together")
 
         self.run_functional_metric_test(
             preds=preds,
             target=target,
             metric_functional=average_precision,
             sk_metric=partial(sk_metric, num_classes=num_classes, average=average),
-            metric_args={"num_classes": num_classes, 'average': average},
+            metric_args={"num_classes": num_classes, "average": average},
         )
 
     def test_average_precision_differentiability(self, preds, sk_metric, target, num_classes, average):
@@ -144,23 +144,27 @@ def test_average_precision(scores, target, expected_score):
 
 
 def test_average_precision_warnings_and_errors():
-    """ Test that the correct errors and warnings gets raised """
+    """Test that the correct errors and warnings gets raised."""
 
     # check average argument
-    with pytest.raises(ValueError, match='Expected argument `average` to be one .*'):
-        AveragePrecision(num_classes=5, average='samples')
+    with pytest.raises(ValueError, match="Expected argument `average` to be one .*"):
+        AveragePrecision(num_classes=5, average="samples")
 
     # check that micro average cannot be used with multilabel input
-    pred = tensor([[0.75, 0.05, 0.05, 0.05, 0.05],
-                   [0.05, 0.75, 0.05, 0.05, 0.05],
-                   [0.05, 0.05, 0.75, 0.05, 0.05],
-                   [0.05, 0.05, 0.05, 0.75, 0.05]])
+    pred = tensor(
+        [
+            [0.75, 0.05, 0.05, 0.05, 0.05],
+            [0.05, 0.75, 0.05, 0.05, 0.05],
+            [0.05, 0.05, 0.75, 0.05, 0.05],
+            [0.05, 0.05, 0.05, 0.75, 0.05],
+        ]
+    )
     target = tensor([0, 1, 3, 2])
-    average_precision = AveragePrecision(num_classes=5, average='micro')
-    with pytest.raises(ValueError, match='Cannot use `micro` average with multi-class input'):
+    average_precision = AveragePrecision(num_classes=5, average="micro")
+    with pytest.raises(ValueError, match="Cannot use `micro` average with multi-class input"):
         average_precision(pred, target)
 
     # check that warning is thrown when average=macro and nan is encoutered in individual scores
-    average_precision = AveragePrecision(num_classes=5, average='macro')
-    with pytest.warns(UserWarning, match='Average precision score for one or more classes was `nan`.*'):
+    average_precision = AveragePrecision(num_classes=5, average="macro")
+    with pytest.warns(UserWarning, match="Average precision score for one or more classes was `nan`.*"):
         average_precision(pred, target)