+++ torchmetrics/functional/classification/average_precision.py
@@ -11,8 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import List, Optional, Sequence, Tuple, Union
 import warnings
+from typing import List, Optional, Sequence, Tuple, Union
 
 import torch
 from torch import Tensor
@@ -28,9 +28,9 @@ def _average_precision_update(
     target: Tensor,
     num_classes: Optional[int] = None,
     pos_label: Optional[int] = None,
-    average: Optional[str] = 'macro',
+    average: Optional[str] = "macro",
 ) -> Tuple[Tensor, Tensor, int, Optional[int]]:
-    """ Format the predictions and target based on the num_classes, pos_label and average parameter 
+    """Format the predictions and target based on the num_classes, pos_label and average parameter
     Args:
         preds: predictions from model (logits or probabilities)
         target: ground truth values
@@ -42,14 +42,14 @@ def _average_precision_update(
         average: reduction method for multi-class or multi-label problems
     """
     preds, target, num_classes, pos_label = _precision_recall_curve_update(preds, target, num_classes, pos_label)
-    if average == 'micro':
+    if average == "micro":
         if preds.ndim == target.ndim:
             # Considering each element of the label indicator matrix as a label
             preds = preds.flatten()
             target = target.flatten()
             num_classes = 1
         else:
-            raise ValueError('Cannot use `micro` average with multi-class input')
+            raise ValueError("Cannot use `micro` average with multi-class input")
 
     return preds, target, num_classes, pos_label
 
@@ -59,7 +59,7 @@ def _average_precision_compute(
     target: Tensor,
     num_classes: int,
     pos_label: Optional[int] = None,
-    average: Optional[str] = 'macro',
+    average: Optional[str] = "macro",
     sample_weights: Optional[Sequence] = None,
 ) -> Union[List[Tensor], Tensor]:
     """Computes the average precision score.
@@ -98,7 +98,7 @@ def _average_precision_compute(
 
     # todo: `sample_weights` is unused
     precision, recall, _ = _precision_recall_curve_compute(preds, target, num_classes, pos_label)
-    if average == 'weighted':
+    if average == "weighted":
         if preds.ndim == target.ndim and target.ndim > 1:
             weights = target.sum(dim=0)
         else:
@@ -113,7 +113,7 @@ def _average_precision_compute_with_precision_recall(
     precision: Tensor,
     recall: Tensor,
     num_classes: int,
-    average: Optional[str] = 'macro',
+    average: Optional[str] = "macro",
     weights: Optional[Tensor] = None,
 ) -> Union[List[Tensor], Tensor]:
     """Computes the average precision score from precision and recall.
@@ -160,12 +160,14 @@ def _average_precision_compute_with_precision_recall(
         res.append(-torch.sum((r[1:] - r[:-1]) * p[:-1]))
 
     # Reduce
-    if average in ('macro', 'weighted'):
+    if average in ("macro", "weighted"):
         res = torch.stack(res)
         if torch.isnan(res).any():
-            warnings.warn('Average precision score for one or more classes was `nan`.'
-                          ' Ignoring these classes in average', UserWarning)
-        if average == 'macro':
+            warnings.warn(
+                "Average precision score for one or more classes was `nan`." " Ignoring these classes in average",
+                UserWarning,
+            )
+        if average == "macro":
             return res[~torch.isnan(res)].mean()
         else:
             weights = torch.ones_like(res) if weights is None else weights
@@ -173,9 +175,8 @@ def _average_precision_compute_with_precision_recall(
     elif average is None:
         return res
     else:
-        allowed_average = ('micro', 'macro', 'weighted', None)
-        raise ValueError(f'Expected argument `average` to be one of {allowed_average}'
-                         f' but got {average}')
+        allowed_average = ("micro", "macro", "weighted", None)
+        raise ValueError(f"Expected argument `average` to be one of {allowed_average}" f" but got {average}")
 
 
 def average_precision(
@@ -183,7 +184,7 @@ def average_precision(
     target: Tensor,
     num_classes: Optional[int] = None,
     pos_label: Optional[int] = None,
-    average: Optional[str] = 'macro',
+    average: Optional[str] = "macro",
     sample_weights: Optional[Sequence] = None,
 ) -> Union[List[Tensor], Tensor]:
     """Computes the average precision score.
@@ -198,7 +199,7 @@ def average_precision(
             this argument should not be set as we iteratively change it in the
             range [0,num_classes-1]
         average:
-            defines the reduction that is applied in the case of multiclass and multilabel input. 
+            defines the reduction that is applied in the case of multiclass and multilabel input.
             Should be one of the following:
 
             - ``'macro'`` [default]: Calculate the metric for each class separately, and average the